Dev improvement - better browser-ability
Build relevant cscope database based on filers for ack, which are much
easier to refine due to exclude (i.e. "not") and double inverted regular
expressions, making it triple inverted with rules in-between. In all
result is fully definable relevant source tree in 5 lines of filers (see
makefile for example).